\name{LongQuest-package}
\alias{LongQuest-package}
\alias{LongQuest}
\docType{package}
\title{
\packageTitle{LongQuest}
}
\description{
\packageDescription{LongQuest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LongQuest}
\packageIndices{LongQuest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LongQuest}

Maintainer: \packageMaintainer{LongQuest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
